+++ internal/codeintel/codenav/internal/lsifstore/lsifstore_documents.go
@@ -60,15 +60,10 @@ func (s *store) GetFullSCIPNameByDescriptor(ctx context.Context, uploadID []int,
 	ctx, _, endObservation := s.operations.getFullSCIPNameByDescriptor.With(ctx, &err, observation.Args{Attrs: []attribute.KeyValue{}})
 	defer endObservation(1, observation.Args{})
 
-	symbolNamesIlike, err := formatSymbolNamesToLikeClause(symbolNames)
-	if err != nil {
-		return nil, err
-	}
-
 	return scanExplodedSymbols(s.db.Query(ctx, sqlf.Sprintf(
 		getFullSCIPNameByDescriptorQuery,
 		pq.Array(uploadID),
-		pq.Array(symbolNamesIlike),
+		pq.Array(formatSymbolNamesToLikeClause(symbolNames)),
 	)))
 }
 
@@ -78,32 +73,34 @@ SELECT DISTINCT
     l2.name AS package_manager,
     l3.name AS package_name,
     l4.name AS package_version,
-    l5.name AS descriptor
+    l5.name AS descriptor_namespace,
+    l6.name AS descriptor_suffix
 -- Initially fuzzy search (see WHERE clause)
-FROM codeintel_scip_symbols_lookup l6
+FROM codeintel_scip_symbols_lookup l7
 
--- Join to symbols table, which will bridge DESCRIPTOR_NO_SUFFIX (syntect) and DESCRIPTOR (precise)
-JOIN codeintel_scip_symbols ss ON ss.upload_id = l6.upload_id AND ss.descriptor_no_suffix_id = l6.id
+-- Join to symbols table, which will bridge DESCRIPTOR_SUFFIX_FUZZY (syntect) and DESCRIPTOR_SUFFIX (precise)
+JOIN codeintel_scip_symbols_lookup_leaves ll ON ll.upload_id = l7.upload_id AND ll.fuzzy_descriptor_suffix_id = l7.id
 
--- Follow parent path from descriptor l5->l4->l3->l2->l1
-JOIN codeintel_scip_symbols_lookup l5 ON l5.upload_id = ss.upload_id AND l5.id = ss.descriptor_id
-JOIN codeintel_scip_symbols_lookup l4 ON l4.upload_id = ss.upload_id AND l4.id = l5.parent_id
-JOIN codeintel_scip_symbols_lookup l3 ON l3.upload_id = ss.upload_id AND l3.id = l4.parent_id
-JOIN codeintel_scip_symbols_lookup l2 ON l2.upload_id = ss.upload_id AND l2.id = l3.parent_id
-JOIN codeintel_scip_symbols_lookup l1 ON l1.upload_id = ss.upload_id AND l1.id = l2.parent_id
+-- Follow parent path from descriptor l6->l5->l4->l3->l2->l1
+JOIN codeintel_scip_symbols_lookup l6 ON l6.upload_id = l6.upload_id AND l6.id = ll.descriptor_suffix_id -- DESCRIPTOR_SUFFIX
+JOIN codeintel_scip_symbols_lookup l5 ON l5.upload_id = l6.upload_id AND l5.id = l6.parent_id            -- DESCRIPTOR_NAMESPACE
+JOIN codeintel_scip_symbols_lookup l4 ON l4.upload_id = l6.upload_id AND l4.id = l5.parent_id            -- PACKAGE_VERSION
+JOIN codeintel_scip_symbols_lookup l3 ON l3.upload_id = l6.upload_id AND l3.id = l4.parent_id            -- PACKAGE_NAME
+JOIN codeintel_scip_symbols_lookup l2 ON l2.upload_id = l6.upload_id AND l2.id = l3.parent_id            -- PACKAGE_MANAGER
+JOIN codeintel_scip_symbols_lookup l1 ON l1.upload_id = l6.upload_id AND l1.id = l2.parent_id            -- SCHEME
 WHERE
-	l6.upload_id = ANY(%s) AND
-	l6.scip_name_type = 'DESCRIPTOR_NO_SUFFIX' AND
-	l6.name ILIKE ANY(%s)
+	l7.upload_id = ANY(%s) AND
+	l7.segment_type = 'DESCRIPTOR_SUFFIX_FUZZY' AND
+	reverse(l7.name) ILIKE ANY(%s)
 `
 
 var scanExplodedSymbols = basestore.NewSliceScanner(func(s dbutil.Scanner) (*symbols.ExplodedSymbol, error) {
 	var n symbols.ExplodedSymbol
-	err := s.Scan(&n.Scheme, &n.PackageManager, &n.PackageName, &n.PackageVersion, &n.Descriptor)
+	err := s.Scan(&n.Scheme, &n.PackageManager, &n.PackageName, &n.PackageVersion, &n.DescriptorNamespace, &n.DescriptorSuffix)
 	return &n, err
 })
 
-func formatSymbolNamesToLikeClause(symbolNames []string) ([]string, error) {
+func formatSymbolNamesToLikeClause(symbolNames []string) []string {
 	trimmedDescriptorMap := make(map[string]struct{}, len(symbolNames))
 	for _, symbolName := range symbolNames {
 		ex, err := symbols.NewExplodedSymbol(symbolName)
@@ -111,18 +108,26 @@ func formatSymbolNamesToLikeClause(symbolNames []string) ([]string, error) {
 			continue
 		}
 
-		// TODO: BIT OF A HACK
-		parts := strings.Split(ex.Descriptor, "/")
-		trimmedDescriptorMap[parts[len(parts)-1]] = struct{}{}
+		trimmedDescriptorMap[ex.FuzzyDescriptorSuffix] = struct{}{}
 	}
 
 	descriptorWildcards := make([]string, 0, len(trimmedDescriptorMap))
 	for symbol := range trimmedDescriptorMap {
 		if symbol != "" {
-			descriptorWildcards = append(descriptorWildcards, "%"+symbol)
+			// TODO - too lenient?
+			descriptorWildcards = append(descriptorWildcards, reverse(symbol)+"%")
 		}
 	}
 	sort.Strings(descriptorWildcards)
 
-	return descriptorWildcards, nil
+	return descriptorWildcards
+}
+
+func reverse(s string) string {
+	b := strings.Builder{}
+	for i := len(s) - 1; i >= 0; i-- {
+		b.WriteByte(s[i])
+	}
+
+	return b.String()
 }