+++ internal/codeintel/uploads/internal/lsifstore/insert_test.go
@@ -2,9 +2,9 @@ package lsifstore
 
 import (
 	"context"
-	"strings"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
 	"github.com/sourcegraph/log/logtest"
 	"github.com/sourcegraph/scip/bindings/go/scip"
 
@@ -179,10 +179,18 @@ func TestInsertDocumentWithSymbols(t *testing.T) {
 
 func TestConstructSymbolLookupTable(t *testing.T) {
 	id := 0
-	symbolNames := []string{
-		"node pnpm pkg1 0.1.0 foo.bar.ident#",
-		"node pnpm pkg1 0.1.1 bar.baz.longerName#",
-		"node pnpm pkg2 0.1.2 baz.bonk.quux#",
+	testCases := []struct {
+		symbolName string
+		parts      []string
+	}{
+		{"node pnpm pkg1 0.1.0 foo.bar.ident#", []string{"node", "pnpm", "pkg1", "0.1.0", "", "foo.bar.ident#"}},
+		{"node pnpm pkg1 0.1.1 bar/`types.ts`/baz/longerName#", []string{"node", "pnpm", "pkg1", "0.1.1", "bar/`types.ts`/baz/", "longerName#"}},
+		{"node pnpm pkg2 0.1.2 baz.bonk.quux#", []string{"node", "pnpm", "pkg2", "0.1.2", "", "baz.bonk.quux#"}},
+	}
+
+	var symbolNames []string
+	for _, testCase := range testCases {
+		symbolNames = append(symbolNames, testCase.symbolName)
 	}
 
 	cache, traverser, err := constructSymbolLookupTable(symbolNames, func() int { v := id; id++; return v })
@@ -191,16 +199,16 @@ func TestConstructSymbolLookupTable(t *testing.T) {
 	}
 
 	type rowType struct {
-		scipNameType string
-		name         string
-		id           int
-		parentID     *int
+		segmentType string
+		name        string
+		id          int
+		parentID    *int
 	}
 	rows := map[int]rowType{}
 
 	// Traverse the tree and build a map of rows by identifier for quick lookup
-	if err := traverser(func(scipNameType, name string, id int, parentID *int) error {
-		rows[id] = rowType{scipNameType, name, id, parentID}
+	if err := traverser(func(segmentType, name string, id int, parentID *int) error {
+		rows[id] = rowType{segmentType, name, id, parentID}
 		return nil
 	}); err != nil {
 		t.Fatalf("unexpected error: %s", err)
@@ -211,9 +219,9 @@ func TestConstructSymbolLookupTable(t *testing.T) {
 	// output match for everything we've inserted. If so we'll consider the tree
 	// to be "well-formed".
 
-	find := func(symbolName string) string {
+	find := func(symbolName string) []string {
 		var parts []string
-		row := rows[cache[symbolName].descriptorID]
+		row := rows[cache[symbolName].descriptorSuffixID]
 		for {
 			// prepend to construct symbol in correct order
 			parts = append([]string{row.name}, parts...)
@@ -224,12 +232,12 @@ func TestConstructSymbolLookupTable(t *testing.T) {
 			row = rows[*row.parentID]
 		}
 
-		return strings.Join(parts, " ")
+		return parts
 	}
 
-	for _, symbolName := range symbolNames {
-		if got := find(symbolName); got != symbolName {
-			t.Errorf("failed to retrieve %q (got %q)", symbolName, got)
+	for _, testCase := range testCases {
+		if diff := cmp.Diff(testCase.parts, find(testCase.symbolName)); diff != "" {
+			t.Errorf("unexpected result (-want +got):\n%s", diff)
 		}
 	}
 }