+++ migrations/codeintel/1689175135_add_enum_type/down.sql
@@ -1,0 +1,1 @@
+DROP TYPE IF EXISTS SymbolNameSegmentType;