+++ migrations/codeintel/1689175157_update_scip_symbol_tables/down.sql
@@ -1,0 +1,3 @@
+DROP TABLE IF EXISTS codeintel_scip_symbols_lookup;
+DROP TABLE IF EXISTS codeintel_scip_symbols_lookup_leaves;
+DROP TABLE IF EXISTS codeintel_scip_symbols_migration_progress;